This function has been removed and it is now provided by the CIMLR package, please refer to https://github.com/danro9685/CIMLR

perform the CIMLR clustering algorithm

CIMLR(X, c, no.dim = NA, k = 10, cores.ratio = 1)

Arguments

X

a list of multi-omic data each of which is an (m x n) data matrix of measurements of cancer patients

c

number of clusters to be estimated over X

no.dim

number of dimensions

k

tuning parameter

cores.ratio

ratio of the number of cores to be used when computing the multi-kernel

Value

clusters the patients based on CIMLR and their similarities

Examples

   
if (FALSE) {
data(GliomasReduced)
CIMLR(X = GliomasReduced$in_X, c = 3, cores.ratio = 0)   
}